requirements.txt
PyMuPDF
google-cloud-texttospeech
prompt-toolkit
tqdm
dotenv
google-genai